--- src/pages/ShipOrder/CreateNewOrder/CreateNewOrder/OrdersByEvents/InitialServices/todo.txt
@@ -1,4 +1,0 @@
-Tarkista tilauksien lisäyksen yhteydessä että lisättävä tilaus ei ole jo lisätty (vältä tuplabuukkauksia)
-Testaukset
-estä poistuminen tilauksien luomis sivulta (?)
-mahdollisesti tilauskien lisääminen välimuistiin
