+++ src/pages/ShipOrder/CreateNewOrder/CreateNewOrder/OrdersByEvents/TemplateSelect.tsx
@@ -1,12 +1,11 @@
-import React, { useState, useEffect } from "react";
+import React, { useState, forwardRef, useEffect } from "react";
 import TextField from "@mui/material/TextField";
-import { parse } from "date-fns";
 import Autocomplete, { createFilterOptions } from "@mui/material/Autocomplete";
 import { OrderTemplate } from "Types";
 import { UseFormSetValue, Control, useWatch } from "react-hook-form";
 import { OrderFormValues } from ".";
 import { getOrderTemplates } from "storage/readAndWriteOrders";
-import { isFriday, isSaturday, isSunday, isWeekend } from "date-fns";
+import { parse, isFriday, isSaturday, isSunday, isWeekend } from "date-fns";
 
 interface TemplateSelectProps{
   control: Control<OrderFormValues>;
@@ -14,7 +13,7 @@ interface TemplateSelectProps{
   ref: React.Ref<HTMLInputElement | null>;
 }
 
-const TemplateSelect = React.forwardRef(({ control, setValue }:TemplateSelectProps, ref) => {
+const TemplateSelect = forwardRef(({ control, setValue }:TemplateSelectProps, ref) => {
   const [ template, setTemplate ] = useState<OrderTemplate|null>();
   const [ templates, setTemplates ] = useState<OrderTemplate[]>([]);
   const date = useWatch({ control, name: "date" });
